+++ apis/s3/src/main/java/org/jclouds/s3/xml/DeleteResultHandler.java
@@ -14,11 +14,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jclouds.aws.s3.xml;
+package org.jclouds.s3.xml;
 
 import static org.jclouds.util.SaxUtils.equalsOrSuffix;
 
-import org.jclouds.aws.s3.domain.DeleteResult;
+import org.jclouds.s3.domain.DeleteResult;
 import org.jclouds.http.functions.ParseSax;
 import org.xml.sax.Attributes;
 import org.xml.sax.SAXException;
