+++ apis/s3/src/test/java/org/jclouds/s3/binders/BindIterableAsPayloadToDeleteRequestTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jclouds.aws.s3.binders;
+package org.jclouds.s3.binders;
 
 import static org.testng.Assert.assertEquals;
