+++ apis/s3/src/test/resources/delete-result.xml
