+++ providers/aws-s3/src/main/java/org/jclouds/aws/s3/AWSS3Client.java
@@ -18,26 +18,10 @@ package org.jclouds.aws.s3;
 
 import static org.jclouds.blobstore.attr.BlobScopes.CONTAINER;
 
-import javax.inject.Named;
-import javax.ws.rs.POST;
-import javax.ws.rs.Path;
-
-import org.jclouds.aws.s3.binders.BindIterableAsPayloadToDeleteRequest;
-import org.jclouds.aws.s3.domain.DeleteResult;
-import org.jclouds.aws.s3.xml.DeleteResultHandler;
 import org.jclouds.blobstore.attr.BlobScope;
-import org.jclouds.rest.annotations.BinderParam;
-import org.jclouds.rest.annotations.EndpointParam;
-import org.jclouds.rest.annotations.ParamValidators;
-import org.jclouds.rest.annotations.QueryParams;
 import org.jclouds.rest.annotations.RequestFilters;
-import org.jclouds.rest.annotations.XMLResponseParser;
-import org.jclouds.s3.Bucket;
 import org.jclouds.s3.S3Client;
-import org.jclouds.s3.binders.BindAsHostPrefixIfConfigured;
 import org.jclouds.s3.filters.RequestAuthorizeSignature;
-import org.jclouds.s3.functions.AssignCorrectHostnameForBucket;
-import org.jclouds.s3.predicates.validators.BucketNameValidator;
 
 /**
  * Provides access to amazon-specific S3 features
@@ -45,32 +29,4 @@ import org.jclouds.s3.predicates.validators.BucketNameValidator;
 @RequestFilters(RequestAuthorizeSignature.class)
 @BlobScope(CONTAINER)
 public interface AWSS3Client extends S3Client {
-
-   /**
-    * The Multi-Object Delete operation enables you to delete multiple objects from a bucket using a 
-    * single HTTP request. If you know the object keys that you want to delete, then this operation 
-    * provides a suitable alternative to sending individual delete requests (see DELETE Object), 
-    * reducing per-request overhead.
-    * 
-    * The Multi-Object Delete request contains a set of up to 1000 keys that you want to delete.
-    * 
-    * If a key does not exist is considered to be deleted. 
-    * 
-    * The Multi-Object Delete operation supports two modes for the response; verbose and quiet.
-    * By default, the operation uses verbose mode in which the response includes the result of
-    * deletion of each key in your request.
-    * 
-    * @param bucketName
-    *           namespace of the objects you are deleting
-    * @param keys
-    *           set of unique keys identifying objects
-    */
-   @Named("DeleteObject")
-   @POST
-   @Path("/")
-   @QueryParams(keys = "delete")
-   @XMLResponseParser(DeleteResultHandler.class)
-   DeleteResult deleteObjects(@Bucket @EndpointParam(parser = AssignCorrectHostnameForBucket.class) @BinderParam(
-         BindAsHostPrefixIfConfigured.class) @ParamValidators(BucketNameValidator.class) String bucketName,
-         @BinderParam(BindIterableAsPayloadToDeleteRequest.class) Iterable<String> keys);
 }
