+++ providers/aws-s3/src/test/java/org/jclouds/aws/s3/AWSS3ClientLiveTest.java
@@ -20,15 +20,10 @@ import static org.jclouds.aws.s3.blobstore.options.AWSS3PutOptions.Builder.stora
 import static org.jclouds.s3.options.ListBucketOptions.Builder.withPrefix;
 import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertNotNull;
-import static org.testng.Assert.assertTrue;
 import static org.testng.Assert.fail;
-
-import java.util.Set;
-import java.util.UUID;
 
 import org.jclouds.aws.AWSResponseException;
 import org.jclouds.aws.domain.Region;
-import org.jclouds.aws.s3.domain.DeleteResult;
 import org.jclouds.blobstore.BlobStore;
 import org.jclouds.blobstore.domain.Blob;
 import org.jclouds.blobstore.domain.StorageMetadata;
@@ -137,34 +132,6 @@ public class AWSS3ClientLiveTest extends S3ClientLiveTest {
          fail("Should had failed because in non-US regions, mixed-case bucket names are invalid.");
       } catch (AWSResponseException e) {
          assertEquals("InvalidBucketName", e.getError().getCode());
-      }
-   }
-   
-   public void testDeleteMultipleObjects() throws InterruptedException {
-      String container = getContainerName();
-      try {
-         ImmutableSet.Builder<String> builder = ImmutableSet.builder();
-         for (int i = 0; i < 5; i++) {
-            String key = UUID.randomUUID().toString();
-            
-            Blob blob = view.getBlobStore().blobBuilder(key).payload("").build();
-            view.getBlobStore().putBlob(container, blob);
-            
-            builder.add(key);
-         }
-
-         Set<String> keys = builder.build();
-         DeleteResult result = getApi().deleteObjects(container, keys);
-
-         assertTrue(result.getDeleted().containsAll(keys));
-         assertEquals(result.getErrors().size(), 0);
-
-         for (String key : keys) {
-            assertConsistencyAwareBlobDoesntExist(container, key);
-         }
-         
-      }  finally {
-         returnContainer(container);
       }
    }
